OPTICAL_FLOW.flow_comp_m_(x/y) units to m/s (#1189)
As stated by @jlecoeur the OPTICAL_FLOW.flow_comp_m_x and flow_comp_m_y must either be velocity units or an integration time unit is required so that flow can be determined from these. Given there is no such integration unit it is very likely that people are assuming m/s OR that it is not being used. Either way, the "best" fix is to clarify m/s.